Create iris_classification.py
Iris flower has three species; setosa, versicolor, and virginica, which differs according to their measurements. Now assume that you have the measurements of the iris flowers according to their species, and here our task was to train a machine learning model that can learn from the measurements of the iris species and classify them.